backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
sqlalchemy==2.0.30
pydantic==2.7.1
python-multipart==0.0.9
pytest==8.2.0
httpx==0.27.0
pytest-asyncio==0.23.6
passlib[bcrypt]==1.7.4
twilio==9.2.3
scikit-learn==1.5.0
numpy==1.26.4
python-dotenv==1.0.1
alembic==1.13.1
